--- a/src/Yields.jl
+++ b/src/Yields.jl
@@ -1,29 +1,9 @@
-function YieldCurve(sg::ScenarioGenerator{N,T,R}; model=FinanceModels.Spline.Cubic(), method=FinanceModels.Fit.Bootstrap()) where {N,T<:InterestRateModel,R}
-    qs = Quotes(sg)
-    FinanceModels.fit(model, qs, method)
-
-end
-
-# function YieldCurve(C::T) where {T<:Correlated}
-#     timestep = first(C.sg).timestep
-#     _, times = __zeros_times(first(C.sg))
-#     _zeros(v) = cumsum(v .* timestep) ./ times
-#     rates = collect(C)
-#     zs = _zeros.(rates)
-#     map(r->Yields.Zero(r,times),zs)
-
-# end
-
-# function __disc_rate_to_fwd(rate,time)
-#     log(rate) / -time
-# end
-
-function Quotes(sg::ScenarioGenerator{N,T,R}) where {N,T<:InterestRateModel,R}
+function YieldCurve(sg::ScenarioGenerator{N,T,R}; model=FinanceModels.Spline.Linear()) where {N,T<:InterestRateModel,R}
     times = sg.timestep:sg.timestep:(sg.endtime+sg.timestep)
     # compute the accumulated discount factor (ZCB price)
-    zeros = cumsum(sg .* sg.timestep) ./ times
-    # the broadcasting versions is about 1/3 fewer allocations
-    # zeros = Iterators.map(/,Iterators.accumulate(+,sg),times)
+    zeros = cumsum(FinanceCore.rate.(sg .* sg.timestep)) ./ times
 
-    return FinanceModels.ZCBYield.(zeros, times)
+    zero_vec = clamp.(zeros, 0.00001, 1)
+    return FinanceModels.Yield.Spline(model, times, zero_vec)
+
 end