--- a/src/interest.jl
+++ b/src/interest.jl
@@ -152,7 +152,9 @@ struct HullWhite{T} <: ShortRateModel
     σ::Float64 # .0168
     curve::T
 end
-# See Yields.jl for HullWhite with a YieldCurve defining theta
+# See FinanceModels.jl for HullWhite with a YieldCurve defining theta
+
+__initial_value(m::HullWhite, timestep) = FinanceCore.forward(m.curve, 0.0, timestep)
 
 # how would HullWhite be constructed if not giving it a curve?
 function nextvalue(M::HullWhite{T}, prior, time, timestep, variate) where {T}
@@ -163,18 +165,18 @@ function nextvalue(M::HullWhite{T}, prior, time, timestep, variate) where {T}
 end
 
 # make this an extension w/ FinanceModels?
-# function θ(M::HullWhite{T}, time, timestep) where {T<:Yields.AbstractYield}
-#     # https://quantpie.co.uk/srm/hull_white_sr.php
-#     # https://quant.stackexchange.com/questions/8724/how-to-calibrate-hull-white-from-zero-curve
-#     # https://mdpi-res.com/d_attachment/mathematics/mathematics-08-01719/article_deploy/mathematics-08-01719-v2.pdf?version=1603181408
-#     a = M.a
-#     f(t) = log(Yields.discount(M.curve, t[1]))
-#     δf = -only(ForwardDiff.hessian(f, [time]))::Float64
-#     f_t = -only(ForwardDiff.gradient(f, [time]))::Float64
+function θ(M::HullWhite{T}, time, timestep) where {T<:Union{FinanceModels.Yield.AbstractYieldModel,FinanceCore.Rate}}
+    # https://quantpie.co.uk/srm/hull_white_sr.php
+    # https://quant.stackexchange.com/questions/8724/how-to-calibrate-hull-white-from-zero-curve
+    # https://mdpi-res.com/d_attachment/mathematics/mathematics-08-01719/article_deploy/mathematics-08-01719-v2.pdf?version=1603181408
+    a = M.a
+    f(t) = log(FinanceCore.discount(M.curve, t[1]))
+    δf = -only(ForwardDiff.hessian(f, [time]))::Float64
+    f_t = -only(ForwardDiff.gradient(f, [time]))::Float64
 
-#     return δf + f_t * a + M.σ^2 / (2 * a) * (1 - exp(-2 * a * time))
+    return δf + f_t * a + M.σ^2 / (2 * a) * (1 - exp(-2 * a * time))
 
-# end
+end
 
 function θ(M::HullWhite{T}, time, timestep) where {T<:Real}
     # https://quantpie.co.uk/srm/hull_white_sr.php
@@ -186,8 +188,3 @@ function θ(M::HullWhite{T}, time, timestep) where {T<:Real}
     return δf + f_t * a + M.σ^2 / (2 * a) * (1 - exp(-2 * a * time))
 
 end
-
-
-# function initial_value(M::HullWhite{T},timestep) where {T}
-#     Yields.forward(M.curve,0,timestep)
-# end