--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -1,12 +1,13 @@
 using EconomicScenarioGenerators
 using Test
-using ActuaryUtilities
-using Yields
+using FinanceCore
+using FinanceModels
 using Distributions
 using Copulas
 using HypothesisTests
 using StatsBase
 using StableRNGs
+using Transducers
 
 include("utils.jl")
 include("interest.jl")