--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -12,5 +12,4 @@ using Transducers
 include("utils.jl")
 include("interest.jl")
 include("equity.jl")
-include("generator.jl")
 include("Correlated.jl")