--- a/test/utils.jl
+++ b/test/utils.jl
@@ -1,19 +1,15 @@
 # http://www.timworrall.com/fin-40008/bscholes.pdf
 
-function price_distribution(m::BlackScholesMerton,T)
+function price_distribution(m::BlackScholesMerton, T)
     r, q, σ = m.r, m.q, m.σ
     return LogNormal(
-        log(m.initial) + (r-q-σ^2/2) * T,
+        log(m.initial) + (r - q - σ^2 / 2) * T,
         sqrt(T) * σ
-        )
+    )
 end
 
-function price_distribution(s::ScenarioGenerator{N,T,R}) where {N,T<:BlackScholesMerton,R}
+function price_distribution(s::ScenarioGenerator{M,N,T,R}) where {M,N,T<:BlackScholesMerton,R}
     m = s.model
     t = s.endtime
-    price_distribution(m,t)
-end
-
-function ratio(x::Vector{T}) where {T}
-    @views x[begin:end-1] ./ x[begin+1:end]
+    price_distribution(m, t)
 end