Todo:

- [x] Figure out performance differences (done: need to `@inline` functions)
- [ ] is it a bad idea to return a tuple instead of an array? 
- [ ] Some docs showing common patterns
- [ ] How should these models be thought of w.r.t. FinanceModels. Are they subtypes of FinanceModels or do they need to be converted/wrapped?
- [ ] HullWhite currently failing - seems the computed derivative in the theta function is slightly different from prior versiosn

## Benchmarking

BM script:

```julia
using EconomicScenarioGenerators
using BenchmarkTools
m = BlackScholesMerton(0.01,0.02,.15,100.)
s = ScenarioGenerator(
    1,  # timestep
    252, # projection horizon
    m,  # model
)

f(s) = map(1:10_000) do i
    collect(s)
end

g(s) = for i in 1:10_000
    collect(s)
end

@benchmark f($s)
@benchmark g($s)


```

Results:

master:

```julia-repl
julia> @benchmark f($s)
BenchmarkTools.Trial: 91 samples with 1 evaluation.
 Range (min … max):  51.027 ms … 93.768 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     53.839 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   55.260 ms ±  6.366 ms  ┊ GC (mean ± σ):  3.38% ± 3.84%

  █▃    ▂                                                      
  ██▄▅▄▄█▄▅▆▃▃▃▃▁▁▃▁▃▁▁▁▁▃▃▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃ ▁
  51 ms           Histogram: frequency by time        87.3 ms <

 Memory estimate: 20.83 MiB, allocs estimate: 10002.

julia> @benchmark g($s)
BenchmarkTools.Trial: 92 samples with 1 evaluation.
 Range (min … max):  50.706 ms … 119.912 ms  ┊ GC (min … max): 0.00% … 6.36%
 Time  (median):     53.374 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   54.652 ms ±   8.720 ms  ┊ GC (mean ± σ):  3.27% ± 3.39%

  █   ▃                                                         
  █▇▄▄██▃▃▃▂▂▁▁▁▁▁▁▁▁▁▁▁▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂ ▁
  50.7 ms         Histogram: frequency by time         98.3 ms <

 Memory estimate: 20.75 MiB, allocs estimate: 10000.
```

This PR:


```julia-repl
julia> @benchmark f($s)
BenchmarkTools.Trial: 38 samples with 1 evaluation.
 Range (min … max):  131.040 ms … 138.651 ms  ┊ GC (min … max): 4.99% … 8.98%
 Time  (median):     132.716 ms               ┊ GC (median):    6.05%
 Time  (mean ± σ):   133.003 ms ±   1.599 ms  ┊ GC (mean ± σ):  6.27% ± 0.87%

     ▃▃▃  ▃▃▃  ▃▃█▃     ▃                                        
  ▇▇▇███▇▁███▇▇████▁▁▇▁▁█▁▇▁▇▇▁▇▇▇▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇▁▁▁▁▁▁▁▁▁▁▇ ▁
  131 ms           Histogram: frequency by time          139 ms <

 Memory estimate: 190.05 MiB, allocs estimate: 5120002.

julia> @benchmark g($s)
BenchmarkTools.Trial: 39 samples with 1 evaluation.
 Range (min … max):  128.833 ms … 148.887 ms  ┊ GC (min … max): 5.35% … 6.43%
 Time  (median):     130.833 ms               ┊ GC (median):    6.00%
 Time  (mean ± σ):   131.538 ms ±   3.763 ms  ┊ GC (mean ± σ):  6.15% ± 0.74%

   █     ▂                                                       
  ██▄█▁▄██▇▄▅▁▁▁▁▁▁▄▄▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄ ▁
  129 ms           Histogram: frequency by time          149 ms <

 Memory estimate: 189.97 MiB, allocs estimate: 5120000.
```

This PR, with `@inline`:

```julia-repl
julia> @benchmark f($s)
BenchmarkTools.Trial: 38 samples with 1 evaluation.
 Range (min … max):  131.040 ms … 138.651 ms  ┊ GC (min … max): 4.99% … 8.98%
 Time  (median):     132.716 ms               ┊ GC (median):    6.05%
 Time  (mean ± σ):   133.003 ms ±   1.599 ms  ┊ GC (mean ± σ):  6.27% ± 0.87%

     ▃▃▃  ▃▃▃  ▃▃█▃     ▃                                        
  ▇▇▇███▇▁███▇▇████▁▁▇▁▁█▁▇▁▇▇▁▇▇▇▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇▁▁▁▁▁▁▁▁▁▁▇ ▁
  131 ms           Histogram: frequency by time          139 ms <

 Memory estimate: 190.05 MiB, allocs estimate: 5120002.

julia> @benchmark g($s)
BenchmarkTools.Trial: 39 samples with 1 evaluation.
 Range (min … max):  128.833 ms … 148.887 ms  ┊ GC (min … max): 5.35% … 6.43%
 Time  (median):     130.833 ms               ┊ GC (median):    6.00%
 Time  (mean ± σ):   131.538 ms ±   3.763 ms  ┊ GC (mean ± σ):  6.15% ± 0.74%

   █     ▂                                                       
  ██▄█▁▄██▇▄▅▁▁▁▁▁▁▄▄▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄ ▁
  129 ms           Histogram: frequency by time          149 ms <

 Memory estimate: 189.97 MiB, allocs estimate: 5120000.

julia> @benchmark f($s)
BenchmarkTools.Trial: 85 samples with 1 evaluation.
 Range (min … max):  54.512 ms … 62.721 ms  ┊ GC (min … max): 0.00% … 9.07%
 Time  (median):     59.115 ms              ┊ GC (median):    5.93%
 Time  (mean ± σ):   58.940 ms ±  1.831 ms  ┊ GC (mean ± σ):  5.47% ± 2.74%

                           ▆   ▁ ▁▁█▆▆▃  ▆     ▃    ▁          
  ▄▁▁▄▄▁▁▇▇▄▇▄▁▁▁▁▁▇▁▄▁▄▄▁▄█▇▄▄█▄██████▇▇█▇▁▄▁▄█▁▄▄▁█▁▁▄▇▁▁▄▇ ▁
  54.5 ms         Histogram: frequency by time        62.5 ms <

 Memory estimate: 74.39 MiB, allocs estimate: 60002.

julia> @benchmark g($s)
BenchmarkTools.Trial: 83 samples with 1 evaluation.
 Range (min … max):  56.602 ms … 85.827 ms  ┊ GC (min … max): 2.25% … 5.89%
 Time  (median):     59.438 ms              ┊ GC (median):    5.72%
 Time  (mean ± σ):   60.236 ms ±  4.046 ms  ┊ GC (mean ± σ):  5.33% ± 1.70%

     ▁ ▁█▁▅▇▁▇                                                 
  ▆▃▆█▅███████▆▅▁▆▅▃▁▁▃▁▁▁▁▁▁▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃ ▁
  56.6 ms         Histogram: frequency by time        75.4 ms <

 Memory estimate: 74.31 MiB, allocs estimate: 60000.
```