@alecloudenback 2023-08-19
Ref #44 for outstanding items